refactor(core): move drawArrowHead from pptx renderer into shared shape module
Extract the DrawingML line-end (arrow head) renderer into
packages/core/src/shape/arrow.ts and export it. The pptx renderer now
imports it instead of defining its own copy, so the docx renderer can
reuse the same head/tail decoration logic. Behaviour is unchanged: the
pptx path used emuToPx(stroke.width, scale) which equals stroke.width *
scale, matching core's applyStroke convention.
